Finish renaming bot
